# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Multi-market investment portfolio tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dividend-calendar:
    get:
      operationId: listDividendEvents
      tags: [dividends]
      summary: List dividend calendar entries
      parameters:
        - name: portfolioId
          in: query
          schema:
            type: ["integer", "null"]
        - name: from
          in: query
          schema:
            type: ["string", "null"]
          description: ISO date
        - name: to
          in: query
          schema:
            type: ["string", "null"]
          description: ISO date
      responses:
        "200":
          description: Dividend events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DividendEvent"
    post:
      operationId: createDividendEvent
      tags: [dividends]
      summary: Create a dividend calendar entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DividendEventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DividendEvent"
  /dividend-calendar/{eventId}:
    patch:
      operationId: updateDividendEvent
      tags: [dividends]
      summary: Update a dividend calendar entry
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DividendEventUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DividendEvent"
    delete:
      operationId: deleteDividendEvent
      tags: [dividends]
      summary: Delete a dividend calendar entry
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── DASHBOARD ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    DividendEvent:
      type: object
      required: [id, portfolioId, holdingId, symbol, name, dividendType, exDate, totalAmount, currency, createdAt]
      properties:
        id:
          type: integer
        portfolioId:
          type: integer
        holdingId:
          type: integer
        symbol:
          type: string
        name:
          type: string
        dividendType:
          type: string
          enum: [ORDINARY, SPECIAL, STOCK, RETURN_OF_CAPITAL]
        exDate:
          type: string
        recordDate:
          type: ["string", "null"]
        paymentDate:
          type: ["string", "null"]
        dividendPerShare:
          type: ["number", "null"]
        quantity:
          type: ["number", "null"]
        grossAmount:
          type: ["number", "null"]
        taxAmount:
          type: ["number", "null"]
        totalAmount:
          type: number
        currency:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    DividendEventInput:
      type: object
      required: [portfolioId, holdingId, dividendType, exDate, totalAmount, currency]
      properties:
        portfolioId:
          type: integer
        holdingId:
          type: integer
        dividendType:
          type: string
          enum: [ORDINARY, SPECIAL, STOCK, RETURN_OF_CAPITAL]
        exDate:
          type: string
        recordDate:
          type: string
        paymentDate:
          type: string
        dividendPerShare:
          type: number
        totalAmount:
          type: number
        currency:
          type: string
        notes:
          type: string
    DividendEventUpdate:
      type: object
      properties:
        dividendType:
          type: string
          enum: [ORDINARY, SPECIAL, STOCK, RETURN_OF_CAPITAL]
        exDate:
          type: string
        recordDate:
          type: ["string", "null"]
        paymentDate:
          type: ["string", "null"]
        dividendPerShare:
          type: ["number", "null"]
        totalAmount:
          type: number
        notes:
          type: ["string", "null"]